refactor(client,ir,types)!: rebuild provider stack — 3-axis codec/transport + IR types
Replace the legacy monolithic Client + ProviderAdapter surface with a 3-axis
provider stack: ModelCodec (wire format) × ModelTransport (auth + endpoint) ×
EndpointShape (URL pattern bridge). Each axis is independently substitutable
and ProviderClient validates compositions at construction time, so combinations
like Vertex-Gemini and Vertex-Anthropic fall out for free.

* IR (src/ir) is now the canonical representation: ModelRequest, ModelResponse,
  ContentPart, Usage, FinishReason, ModelStreamChunk, ProviderCapabilities,
  ModelWarning. Adds ResponseFormat (Text/JsonObject/JsonSchema) and
  ContentPart::ToolResult.tool_name so Gemini's functionResponse round-trips.

* 5 codecs (anthropic-messages, openai-chat, openai-responses, gemini-generate,
  bedrock-converse) and 4 transports (direct, vertex, bedrock, foundry).
  ModelTransport gains classify_error(status, body) for vendor-specific HTTP
  failure patterns — moves Vertex quota / Bedrock throttling / Foundry Entra
  out of provider_client.rs (OCP). DirectTransport gains
  with_credential_provider + refresh() for OAuth token rotation.

* Capability declarations are now honest: every codec emits the wire field for
  Native capabilities (response_format → text.format / response_format /
  responseSchema) or a CapabilityEmulated warning for Emulated. Bedrock
  Converse cachePoint is wired through apply_cache_points so prompt_caching:
  Native is true.

* types/ keeps only the local executor contract (ToolError, ToolInput,
  ToolOutput, ToolResult, ToolOutputBlock, ToolSpec, ModelUsage, UsageProvider,
  ServerToolUse). ToolDefinition was renamed to ToolSpec so ir::ToolDefinition
  is the only ToolDefinition in the crate; strict and defer_loading collapsed
  to bool. From<&ToolSpec>/From<ToolSpec> for ir::ToolDefinition.

* Pricing table populated for current models — Claude 4-family, GPT-5 family,
  Gemini 2.5 family — with a generic BRANCHFORGE_PRICING_<MODEL>_INPUT
  override.

Removes:
  - src/client/{adapter/*, batch, files, gateway, messages/*, network,
    provider_profile, recovery, streaming}.rs
  - src/types/{message, response, content/*, document, citations, search}.rs
  - src/ir/compat.rs
  - src/types/tool/server.rs (moved to src/agent/server_tools.rs)

+++ src/observability/spans.rs
@@ -109,12 +109,12 @@ impl ApiCallSpan {
         }
     }
 
-    pub fn record_usage(&self, input_tokens: u32, output_tokens: u32) {
+    pub fn record_usage(&self, input_tokens: u64, output_tokens: u64) {
         self.span.record("input_tokens", input_tokens);
         self.span.record("output_tokens", output_tokens);
     }
 
-    pub fn record_cache(&self, read_tokens: u32, creation_tokens: u32) {
+    pub fn record_cache(&self, read_tokens: u64, creation_tokens: u64) {
         self.span.record("cache_read_tokens", read_tokens);
         self.span.record("cache_creation_tokens", creation_tokens);
     }
